fix: Add a loading button between "Backup" request and flagship answer
Backup page behavior is entirely controlled by backup state which is
handled by flagship app. Currently, flagship app can respond slowly
to backup requests when there are a lot of photos.

For example, when you click on "Backup", you can have seconds before
the button is disabled and the user sees the backup as started. So
here we fixed this UX issue by adding a loading button from when the
button is clicked to when the flagship app answer arrives.

This UX fix may not be necessary when the backup will be
optimized enough.