fix: Create album button blank page
In 5587dcaf4c5eee6af27de07cb476c6fc25de8cde, we added params to
ALBUMS_MUTATIONS with an extra function in CreateAlbumPicker leading
to the following error "Functions are not valid as a React child.
This may happen if you return a Component instead of <Component />
from render. Or maybe you meant to call this function rather than
return it."